+++ spi/src/main/java/org/jboss/pnc/spi/datastore/predicates/BuildRecordPredicates.java
@@ -279,7 +279,7 @@ public class BuildRecordPredicates {
         }
     }
 
-    public static Predicate<BuildRecord> withBuildConfigSetRecordId(Integer buildConfigSetRecordId) {
+    public static Predicate<BuildRecord> withBuildConfigSetRecordId(Long buildConfigSetRecordId) {
         return (root, query, cb) -> {
             Join<BuildRecord, BuildConfigSetRecord> joinedConfigSetRecord = root
                     .join(BuildRecord_.buildConfigSetRecord);
